feat: 채팅기능 구현, 상태 기능 구현, security refactoring

See merge request dign552/daejeon_5_13_finalproject_travelquest!1

+++ backend/src/main/java/com/ssafy/travelquest/domain/user/repository/UserRepository.java
@@ -9,7 +9,7 @@ import com.ssafy.travelquest.domain.user.entity.User;
 
 @Mapper
 public interface UserRepository {
-    User findById(int no);
+    User findById(Long no);
     User findByUserId(String userId);
     int insert(User user);
     void update(User user);
